moved Setk to arrow-core-data
